txt requirements.txt
instaloader==4.9.5
pytesseract==0.3.10
Pillow==9.5.0